+++ forge/src/main/java/owmii/powah/forge/block/ForgeCableTile.java
@@ -5,7 +5,6 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.energy.CapabilityEnergy;
 import owmii.powah.block.Tier;
@@ -19,7 +18,7 @@ public class ForgeCableTile extends CableTile {
 	}
 
 	@Override
-	public long receiveEnergy(int maxReceive, boolean simulate, @Nullable Direction direction) {
+	public long receiveEnergy(long maxReceive, boolean simulate, @Nullable Direction direction) {
 		if (this.level == null || isRemote() || direction == null || !checkRedstone() || !canReceiveEnergy(direction))
 			return 0;
 		long received = 0;
